pipeline/requirements.txt
yt-dlp
edge-tts
openai
python-dotenv
pydub
openai-whisper
torch
